refactoring funzioni game_utils
